Refactor check_shapes into multiple files.
