+++ tests/gpflow/experimental/check_shapes/test_check_shapes.py
@@ -11,31 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""
-Unit tests for tool for checking the shapes of function using tf Tensors.
-"""
-from abc import ABC, abstractmethod
 from dataclasses import dataclass
-from typing import Callable, Optional, Tuple
-from unittest.mock import MagicMock
+from typing import Tuple
 
-import numpy as np
 import pytest
-import tensorflow as tf
 
-from gpflow.experimental.check_shapes import (
-    ArgumentReferenceError,
-    Shaped,
-    ShapeMismatchError,
-    check_shapes,
-    inherit_check_shapes,
-)
+from gpflow.base import TensorType
+from gpflow.experimental.check_shapes import ShapeMismatchError, check_shapes
 
-
-def t(*shape: Optional[int]) -> Shaped:
-    mock_tensor = MagicMock()
-    mock_tensor.shape = shape
-    return mock_tensor
+from .utils import t
 
 
 def test_check_shapes__constant() -> None:
@@ -44,7 +28,7 @@ def test_check_shapes__constant() -> None:
         ("b", [2, 4]),
         ("return", [3, 4]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(3, 4)
 
     # Don't crash...
@@ -60,7 +44,7 @@ def test_check_shapes__constant__bad_input() -> None:
         ("b", [2, 4]),
         ("return", [3, 4]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(3, 4)
 
     with pytest.raises(ShapeMismatchError):
@@ -73,7 +57,7 @@ def test_check_shapes__constant__bad_return() -> None:
         ("b", [2, 4]),
         ("return", [3, 4]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(3, 4 + 1)
 
     with pytest.raises(ShapeMismatchError):
@@ -86,7 +70,7 @@ def test_check_shapes__var_dim() -> None:
         ("b", ["d1", "d3"]),
         ("return", ["d2", "d3"]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(3, 4)
 
     # Don't crash...
@@ -102,7 +86,7 @@ def test_check_shapes__var_dim__bad_input() -> None:
         ("b", ["d1", "d3"]),
         ("return", ["d2", "d3"]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(3, 4)
 
     with pytest.raises(ShapeMismatchError):
@@ -115,7 +99,7 @@ def test_check_shapes__var_dim__bad_return() -> None:
         ("b", ["d1", "d3"]),
         ("return", ["d2", "d3"]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(3, 4 + 1)
 
     with pytest.raises(ShapeMismatchError):
@@ -128,7 +112,7 @@ def test_check_shapes__var_rank() -> None:
         ("b", ["ds...", "d1"]),
         ("return", ["ds...", "d1", "d2"]),
     )
-    def f(a: Shaped, b: Shaped, leading_dims: int) -> Shaped:
+    def f(a: TensorType, b: TensorType, leading_dims: int) -> TensorType:
         output_shape = leading_dims * (2,) + (3, 4)
         return t(*output_shape)
 
@@ -148,7 +132,7 @@ def test_check_shapes__var_rank__bad_input() -> None:
         ("b", ["ds...", "d1"]),
         ("return", ["ds...", "d1", "d2"]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(1, 2, 3)
 
     with pytest.raises(ShapeMismatchError):
@@ -161,7 +145,7 @@ def test_check_shapes__var_rank__bad_return() -> None:
         ("b", ["ds...", "d1"]),
         ("return", ["ds...", "d1", "d2"]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(1, 2 + 1, 3)
 
     with pytest.raises(ShapeMismatchError):
@@ -174,7 +158,7 @@ def test_check_shapes__scalar() -> None:
         ("b", []),
         ("return", []),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t()
 
     f(t(), t())  # Don't crash...
@@ -186,7 +170,7 @@ def test_check_shapes__scalar__bad_input() -> None:
         ("b", []),
         ("return", []),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t()
 
     with pytest.raises(ShapeMismatchError):
@@ -199,7 +183,7 @@ def test_check_shapes__scalar__bad_return() -> None:
         ("b", []),
         ("return", []),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(1, 1)
 
     with pytest.raises(ShapeMismatchError):
@@ -212,45 +196,21 @@ def test_check_shapes__invalid_argument() -> None:
         ("b", [2, 4]),
         ("return", [3, 4]),
     )
-    def f(a: Shaped, b: Shaped) -> Shaped:
+    def f(a: TensorType, b: TensorType) -> TensorType:
         return t(3, 4)
 
     with pytest.raises(TypeError):
         f(c=t(2, 3))
 
 
-def test_check_shapes__invalid_argument_name() -> None:
-    with pytest.raises(AssertionError):
-
-        @check_shapes(
-            ("a", [2, 3]),
-            ("#$%#$", [2, 4]),
-            ("return", [3, 4]),
-        )
-        def f(a: Shaped, b: Shaped) -> Shaped:
-            return t(3, 4)
-
-
-def test_check_shapes__invalid_nested_argument_name() -> None:
-    with pytest.raises(AssertionError):
-
-        @check_shapes(
-            ("a", [2, 3]),
-            ("b#$%#$", [2, 4]),
-            ("return", [3, 4]),
-        )
-        def f(a: Shaped, b: Shaped) -> Shaped:
-            return t(3, 4)
-
-
 def test_check_shapes__argument_refs() -> None:
     @dataclass
     class Input:
-        ins: Tuple[Shaped, Shaped]
+        ins: Tuple[TensorType, TensorType]
 
     @dataclass
     class Output:
-        out: Shaped
+        out: TensorType
 
     @check_shapes(
         ("x.ins[0]", ["a_batch...", 1]),
@@ -268,266 +228,3 @@ def test_check_shapes__argument_refs() -> None:
 
     with pytest.raises(ShapeMismatchError):
         f(Input(ins=(t(2, 1, 1), t(1, 1, 2))))
-
-
-def test_check_shapes__argument_refs__bad_attribute() -> None:
-    @dataclass
-    class Input:
-        ins: Shaped
-
-    @check_shapes(
-        ("x.outs", ["a_batch...", 1]),
-    )
-    def f(x: Input) -> None:
-        pass
-
-    with pytest.raises(ArgumentReferenceError):
-        f(Input(ins=t(3, 2, 1)))
-
-
-def test_check_shapes__argument_refs__bad_index() -> None:
-    @check_shapes(
-        ("x[2]", ["a_batch...", 1]),
-    )
-    def f(x: Tuple[Shaped, Shaped]) -> None:
-        pass
-
-    with pytest.raises(ArgumentReferenceError):
-        f((t(3, 2, 1), t(2, 3, 1)))
-
-
-def test_check_shapes__numpy() -> None:
-    @check_shapes(
-        ("a", ["d1", "d2"]),
-        ("b", ["d1", "d3"]),
-        ("return", ["d2", "d3"]),
-    )
-    def f(a: np.ndarray, b: np.ndarray) -> np.ndarray:
-        return np.zeros((3, 4))
-
-    f(np.zeros((2, 3)), np.zeros((2, 4)))  # Don't crash...
-
-
-def test_check_shapes__tensorflow() -> None:
-    @check_shapes(
-        ("a", ["d1", "d2"]),
-        ("b", ["d1", "d3"]),
-        ("return", ["d2", "d3"]),
-    )
-    def f(a: tf.Tensor, b: tf.Tensor) -> tf.Tensor:
-        return tf.zeros((3, 4))
-
-    f(tf.zeros((2, 3)), tf.zeros((2, 4)))  # Don't crash...
-
-
-@pytest.mark.parametrize(
-    "f_wrapper,loss_wrapper",
-    [
-        (
-            lambda x: x,
-            lambda x: x,
-        ),
-        (
-            tf.function,
-            tf.function,
-        ),
-        (
-            tf.function(input_signature=[tf.TensorSpec(shape=[None], dtype=tf.float64)]),
-            tf.function(input_signature=[]),
-        ),
-        (
-            tf.function(input_signature=[tf.TensorSpec(shape=None, dtype=tf.float64)]),
-            tf.function(input_signature=[]),
-        ),
-        (
-            tf.function(experimental_relax_shapes=True),
-            tf.function(experimental_relax_shapes=True),
-        ),
-    ],
-)
-def test_check_shapes__tensorflow_compilation(
-    f_wrapper: Callable[[tf.Tensor], tf.Tensor], loss_wrapper: Callable[[], tf.Tensor]
-) -> None:
-    target = 0.5
-
-    @f_wrapper
-    @check_shapes(
-        ("x", ["n"]),
-        ("return", ["n"]),
-    )
-    def f(x: tf.Tensor) -> Tuple[tf.Tensor]:
-        return (x - target) ** 2
-
-    v = tf.Variable(np.linspace(0.0, 1.0))
-
-    @loss_wrapper
-    @check_shapes(
-        ("return", [1]),
-    )
-    def loss() -> tf.Tensor:
-        # keepdims is just to add an extra dimension to make the check more interesting.
-        return tf.reduce_sum(f(v), keepdims=True)
-
-    optimiser = tf.keras.optimizers.SGD(learning_rate=0.25)
-    for _ in range(10):
-        optimiser.minimize(loss, var_list=[v])
-
-    np.testing.assert_allclose(target, v.numpy(), atol=0.01)
-
-
-def test_inherit_check_shapes__defined_in_super_class() -> None:
-    class SuperClass(ABC):
-        @abstractmethod
-        @check_shapes(
-            ("a", [4]),
-            ("return", [1]),
-        )
-        def f(self, a: Shaped) -> Shaped:
-            pass
-
-    class MiddleClass(SuperClass):
-        pass
-
-    class SubClass(MiddleClass):
-        @inherit_check_shapes
-        def f(self, a: Shaped) -> Shaped:
-            return t(1)
-
-    sub = SubClass()
-    sub.f(t(4))  # Don't crash...
-
-    with pytest.raises(ShapeMismatchError):
-        sub.f(t(5))
-
-
-def test_inherit_check_shapes__overridden_with_checks() -> None:
-    class SuperClass(ABC):
-        @abstractmethod
-        @check_shapes(
-            ("a", [4]),
-            ("return", [1]),
-        )
-        def f(self, a: Shaped) -> Shaped:
-            pass
-
-    class MiddleClass(SuperClass):
-        @inherit_check_shapes
-        def f(self, a: Shaped) -> Shaped:
-            return t(2)
-
-    class SubClass(MiddleClass):
-        @inherit_check_shapes
-        def f(self, a: Shaped) -> Shaped:
-            return t(1)
-
-    sub = SubClass()
-    sub.f(t(4))  # Don't crash...
-
-    with pytest.raises(ShapeMismatchError):
-        sub.f(t(5))
-
-
-def test_inherit_check_shapes__overridden_without_checks() -> None:
-    class SuperClass(ABC):
-        @abstractmethod
-        @check_shapes(
-            ("a", [4]),
-            ("return", [1]),
-        )
-        def f(self, a: Shaped) -> Shaped:
-            pass
-
-    class MiddleClass(SuperClass):
-        def f(self, a: Shaped) -> Shaped:
-            return t(2)
-
-    class SubClass(MiddleClass):
-        @inherit_check_shapes
-        def f(self, a: Shaped) -> Shaped:
-            return t(1)
-
-    sub = SubClass()
-    sub.f(t(4))  # Don't crash...
-
-    with pytest.raises(ShapeMismatchError):
-        sub.f(t(5))
-
-
-def test_inherit_check_shapes__defined_in_middle_class() -> None:
-    class SuperClass(ABC):
-        pass
-
-    class MiddleClass(SuperClass):
-        @abstractmethod
-        @check_shapes(
-            ("a", [4]),
-            ("return", [1]),
-        )
-        def f(self, a: Shaped) -> Shaped:
-            pass
-
-    class SubClass(MiddleClass):
-        @inherit_check_shapes
-        def f(self, a: Shaped) -> Shaped:
-            return t(1)
-
-    sub = SubClass()
-    sub.f(t(4))  # Don't crash...
-
-    with pytest.raises(ShapeMismatchError):
-        sub.f(t(5))
-
-
-# In multiple inheratiance
-# Missing
-
-
-def test_inherit_check_shapes__multiple_inheritance() -> None:
-    class Left(ABC):
-        @abstractmethod
-        @check_shapes(
-            ("a", [4]),
-            ("return", [1]),
-        )
-        def f(self, a: Shaped) -> Shaped:
-            pass
-
-    class Right(ABC):
-        @abstractmethod
-        @check_shapes(
-            ("a", [5]),
-            ("return", [1]),
-        )
-        def g(self, a: Shaped) -> Shaped:
-            pass
-
-    class SubClass(Left, Right):
-        @inherit_check_shapes
-        def f(self, a: Shaped) -> Shaped:
-            return t(1)
-
-        @inherit_check_shapes
-        def g(self, a: Shaped) -> Shaped:
-            return t(1)
-
-    sub = SubClass()
-    sub.f(t(4))  # Don't crash...
-    sub.g(t(5))  # Don't crash...
-
-    with pytest.raises(ShapeMismatchError):
-        sub.f(t(5))
-
-    with pytest.raises(ShapeMismatchError):
-        sub.g(t(4))
-
-
-def test_inherit_check_shapes__undefined() -> None:
-    class SuperClass(ABC):
-        pass
-
-    with pytest.raises(RuntimeError):
-
-        class SubClass(SuperClass):
-            @inherit_check_shapes
-            def f(self, a: Shaped) -> Shaped:
-                return t(1)
